# internal/adapter/persistence/sql/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.26.0

package sql

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Todo struct {
	ID        pgtype.UUID
	CreatedAt pgtype.Timestamp
	UpdatedAt pgtype.Timestamp
	Title     string
	Content   string
}